Add movePlayer function. You can move player's character.
