--- README.md
@@ -1,5 +1,0 @@
-# Avoid_Virus_Game
-
-This is the first step of project.
-
-This branch includes initial screen functions.
